+++ R/compare-heaps.R
@@ -21,9 +21,13 @@
 #' @export
 #' @examples
 #' graph <- weight_streetnet (hampi)
+#' \dontrun{
 #' compare_heaps (graph, nverts = 1000, replications = 1)
-compare_heaps <- function(graph, nverts = 100, replications = 2)
-{
+#' }
+compare_heaps <- function (graph,
+                           nverts = 100,
+                           replications = 2) {
+
     requireNamespace ("rbenchmark")
     requireNamespace ("igraph")