+++ tests/testthat/test-flows.R
@@ -1,7 +1,6 @@
 context("dodgr_flows")
 
-test_all <- (identical (Sys.getenv ("MPADGE_LOCAL"), "true") |
-             identical (Sys.getenv ("TRAVIS"), "true"))
+test_all <- (identical (Sys.getenv ("MPADGE_LOCAL"), "true"))
 
 testthat::skip_on_cran ()
 
@@ -15,7 +14,7 @@ test_that("flows aggregate", {
     flows <- matrix (10 * runif (length (from) * length (to)),
                      nrow = length (from))
 
-    expect_message (graph2 <- dodgr_flows_aggregate (graph, from = from, 
+    expect_message (graph2 <- dodgr_flows_aggregate (graph, from = from,
                                                      to = to, flows = flows,
                                                      quiet = FALSE),
                     "Aggregating flows ...")
@@ -25,21 +24,24 @@ test_that("flows aggregate", {
 
     flows [1, 2] <- NA
     graph3 <- dodgr_flows_aggregate (graph, from = from, to = to, flows = flows)
-    if (test_all)
-        #expect_true (max (graph3$flow) <= max (graph2$flow))
+    #if (test_all)
+    #    expect_true (max (graph3$flow) <= max (graph2$flow))
 
     graph4 <- dodgr_flows_aggregate (graph, from = from, to = to, flows = flows,
                                      contract = TRUE)
     # this test is not longer true with aggregated flows normalised via #121:
-    if (test_all)
-        #expect_true (all ((graph4$flow - graph3$flow) < 1e-3))
+    #if (test_all)
+    #    expect_true (all ((graph4$flow - graph3$flow) < 1e-3))
 
     expect_warning (graph4 <- dodgr_flows_aggregate (graph3, from = from,
                                                      to = to, flows = flows),
-                    "graph already has a 'flow' column; this will be overwritten")
+                "graph already has a 'flow' column; this will be overwritten")
 
     flowsv <- as.vector (flows)
-    graph5 <- dodgr_flows_aggregate (graph, from = from, to = to, flows = flowsv)
+    graph5 <- dodgr_flows_aggregate (graph,
+                                     from = from,
+                                     to = to,
+                                     flows = flowsv)
     expect_equal (graph5$flow, graph4$flow)
 })
 
@@ -84,8 +86,10 @@ test_that ("flows disperse", {
     expect_true (all (c ("flow1", "flow2") %in% names (graph3b)))
     expect_equal (graph3a$flow, graph3b$flow1)
 
-    expect_silent (graph4 <- dodgr_flows_disperse (graph, from = from,
-                                                   dens = dens, contract = TRUE))
+    expect_silent (graph4 <- dodgr_flows_disperse (graph,
+                                                   from = from,
+                                                   dens = dens,
+                                                   contract = TRUE))
     # Dispersed flows calculated on contracted graph should **NOT** equal those
     # calculated on full graph
     if (test_all) # fails on CRAN
@@ -125,8 +129,12 @@ test_that ("flows_si", {
 
 
                # calculation via flows_si:
-               netf_si <- dodgr_flows_si (graph, from = from, to = to, k = k,
-                                          dens_from = dens_from, dens_to = dens_to)
+               netf_si <- dodgr_flows_si (graph,
+                                          from = from,
+                                          to = to,
+                                          k = k,
+                                          dens_from = dens_from,
+                                          dens_to = dens_to)
                expect_identical (dim (netf), dim (netf_si))
                expect_identical (names (netf), names (netf_si))
                r2 <- cor (netf$flow, netf_si$flow) ^ 2
@@ -143,8 +151,9 @@ test_that ("flowmap", {
                      nrow = length (from))
     graph <- dodgr_flows_aggregate (graph, from = from, to = to, flows = flows)
     graph_undir <- merge_directed_graph (graph)
-    if (nrow (graph_undir) > 0)
-    {
+
+    if (nrow (graph_undir) > 0) {
+
         # just test that is produces a plot
         png (filename = "junk.png")
         expect_silent (dodgr_flowmap (graph_undir))
@@ -162,4 +171,3 @@ test_that ("flowmap", {
     expect_error (graph_undir <- merge_directed_graph (graph),
                   "col_names \\[flow\\] do not match columns in graph")
 })
-